DLG/scaffold/requirements.txt
# Core dependencies for DLG attack on SCAFFOLD
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.5.0

# Optional dependencies for advanced features
# scipy>=1.10.0  # For advanced metrics
# tqdm>=4.65.0   # For progress bars

